Fix StackOverflowError in extra-long Text Block by using the possessive quantifier to avoid backtracking
Possessive quantifiers are a way to prevent the regex engine from trying all permutations. Because it is impossible that there can be another match with a `-` at the end of the input (it's all or nothing), it is pointless backtracking to find other possible matches.

See https://www.regular-expressions.info/possessive.html